Consolidate components
