+++ client/app/queue/CaseListTable.jsx
@@ -1,12 +1,14 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
 import { css } from 'glamor';
 
 import CaseDetailsLink from './CaseDetailsLink';
 import DocketTypeBadge from './components/DocketTypeBadge';
 import Table from '../components/Table';
 import { COLORS } from '../constants/AppConstants';
+import { clearCaseListSearch } from './CaseList/CaseListActions';
 
 import { DateString } from '../util/DateUtil';
 import { renderAppealType } from './utils';
@@ -31,6 +33,8 @@ const labelForLocation = (locationCode, userId) => {
 };
 
 class CaseListTable extends React.PureComponent {
+  componentWillUnmount = () => this.props.clearCaseListSearch();
+
   getKeyForRow = (rowNumber, object) => object.id;
 
   getColumns = () => [
@@ -84,5 +88,8 @@ const mapStateToProps = (state) => ({
   userCssId: state.ui.userCssId
 });
 
-export default connect(mapStateToProps)(CaseListTable);
+const mapDispatchToProps = (dispatch) => bindActionCreators({
+  clearCaseListSearch
+}, dispatch);
 
+export default connect(mapStateToProps, mapDispatchToProps)(CaseListTable);
