+++ client/app/queue/CaseListView.jsx
@@ -3,30 +3,40 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
+import { css } from 'glamor';
 
 import AppSegment from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/AppSegment';
+import Link from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/Link';
 
 import ApiUtil from '../util/ApiUtil';
 import LoadingDataDisplay from '../components/LoadingDataDisplay';
-import { LOGO_COLORS } from '../constants/AppConstants';
+import {
+  COLORS,
+  LOGO_COLORS
+} from '../constants/AppConstants';
 import CaseListSearch from './CaseListSearch';
 import CaseListTable from './CaseListTable';
 import { fullWidth } from './constants';
 
-import { clearCaseListSearch, onReceiveAppealsUsingVeteranId } from './CaseList/CaseListActions';
+import { onReceiveAppealsUsingVeteranId } from './CaseList/CaseListActions';
 import { appealsByCaseflowVeteranId } from './selectors';
 
 import COPY from '../../COPY.json';
 
-class CaseListView extends React.PureComponent {
-  componentWillUnmount = () => this.props.clearCaseListSearch();
+const horizontalRuleStyling = css({
+  border: 0,
+  borderTop: `1px solid ${COLORS.GREY_LIGHT}`,
+  marginTop: '5rem',
+  marginBottom: '5rem'
+});
 
+class CaseListView extends React.PureComponent {
   createLoadPromise = () => {
-    if (this.props.appeals.length) {
+    const caseflowVeteranId = this.props.caseflowVeteranId;
+
+    if (this.props.appeals.length || !caseflowVeteranId) {
       return Promise.resolve();
     }
-
-    const caseflowVeteranId = this.props.caseflowVeteranId;
 
     return ApiUtil.get(`/cases/${caseflowVeteranId}`).
       then((response) => {
@@ -36,11 +46,21 @@ class CaseListView extends React.PureComponent {
       });
   };
 
+  searchPageHeading = () => <React.Fragment>
+    <h1 className="cf-push-left" {...fullWidth}>{COPY.CASE_SEARCH_HOME_PAGE_HEADING}</h1>
+    <p>{COPY.CASE_SEARCH_INPUT_INSTRUCTION}</p>
+    <CaseListSearch elementId="searchBarEmptyList" />
+  </React.Fragment>;
+
   caseListTable = () => {
     const appealsCount = this.props.appeals.length;
 
     if (!appealsCount) {
-      return null;
+      return <div>
+        {this.searchPageHeading()}
+        <hr {...horizontalRuleStyling} />
+        <p><Link href="/help">Caseflow Help</Link></p>
+      </div>;
     }
 
     // Using the first appeal in the list to get the Veteran's name and ID. We expect that data to be
@@ -50,9 +70,7 @@ class CaseListView extends React.PureComponent {
         “${firstAppeal.veteranFullName} (${firstAppeal.veteranFileNumber})”`;
 
     return <div>
-      <h1 className="cf-push-left" {...fullWidth}>{COPY.CASE_SEARCH_HOME_PAGE_HEADING}</h1>
-      <p>{COPY.CASE_SEARCH_INPUT_INSTRUCTION}</p>
-      <CaseListSearch elementId="searchBarEmptyList" />
+      {this.searchPageHeading()}
       <br /><br />
       <h2 className="cf-push-left" {...fullWidth}>{heading}</h2>
       <CaseListTable appeals={this.props.appeals} />
@@ -93,7 +111,6 @@ const mapStateToProps = (state, ownProps) => ({
 });
 
 const mapDispatchToProps = (dispatch) => bindActionCreators({
-  clearCaseListSearch,
   onReceiveAppealsUsingVeteranId
 }, dispatch);
